+++ crates/client/src/client.rs
@@ -21,6 +21,7 @@ pub use crate::{
 };
 use anyhow::anyhow;
 pub use entropy_protocol::{sign_and_encrypt::EncryptedSignedMessage, KeyParams};
+use parity_scale_codec::Decode;
 use rand::Rng;
 use std::str::FromStr;
 pub use synedrion::KeyShare;
@@ -443,4 +444,22 @@ pub async fn request_attestation(
     attestee: &sr25519::Pair,
 ) -> Result<[u8; 32], ClientError> {
     Ok(user::request_attestation(api, rpc, attestee).await?)
+}
+
+/// Get oracle data headings
+/// This is useful for program developers to know what oracle data is available
+pub async fn get_oracle_headings(
+    api: &OnlineClient<EntropyConfig>,
+    _rpc: &LegacyRpcMethods<EntropyConfig>,
+) -> Result<Vec<String>, ClientError> {
+    let storage_address = entropy::storage().oracle().oracle_data_iter();
+    let mut iter = api.storage().at_latest().await?.iter(storage_address).await?;
+    let mut headings = Vec::new();
+    while let Some(Ok(kv)) = iter.next().await {
+        // Key is: storage_address || 128 bit hash || key
+        let mut input = &kv.key_bytes[32 + 16 + 1..];
+        let heading = String::decode(&mut input)?;
+        headings.push(heading);
+    }
+    Ok(headings)
 }